pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.0",
    "livekit-api>=1.0.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.0",
]
